Move TurboSnap files to lib
