+++ node-src/lib/turbosnap/compareBaseline.test.ts
@@ -2,9 +2,9 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import { describe, expect, it } from 'vitest';
 
+import TestLogger from '../testLogger';
 import { compareBaseline } from './compareBaseline';
 import { getDependencies } from './getDependencies';
-import TestLogger from './testLogger';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
@@ -17,12 +17,12 @@ describe('compareBaseline', () => {
   it('finds changed dependency names', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/react-async-10'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/react-async-10'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/react-async-9'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/react-async-9'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
@@ -33,12 +33,12 @@ describe('compareBaseline', () => {
   it('finds added dependency names', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/react-async-9'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/react-async-9'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/plain'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/plain'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
@@ -49,12 +49,12 @@ describe('compareBaseline', () => {
   it('finds removed dependency names', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/plain'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/plain'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/react-async-9'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/react-async-9'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
@@ -65,12 +65,12 @@ describe('compareBaseline', () => {
   it('finds nothing given identical files', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/plain'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/plain'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/plain'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/plain'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
@@ -81,12 +81,12 @@ describe('compareBaseline', () => {
   it('runs the manifest check on yarn berry lock files successfully', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/berry'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/berry'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/berry'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/berry'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
@@ -97,12 +97,12 @@ describe('compareBaseline', () => {
   it('does not find yarn berry changed dependency name for set resolution', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/berry'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/berry'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/berry-chalk'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/berry-chalk'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
@@ -113,12 +113,12 @@ describe('compareBaseline', () => {
   it('finds yarn berry dependency change name', async () => {
     const ctx = getContext();
     const headDependencies = await getDependencies(ctx, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/berry'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/berry'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
     const baselineChanges = await compareBaseline(ctx, headDependencies, {
-      rootPath: path.join(__dirname, '../__mocks__/dependencyChanges/berry-chalk'),
+      rootPath: path.join(__dirname, '../../__mocks__/dependencyChanges/berry-chalk'),
       manifestPath: 'package.json',
       lockfilePath: 'yarn.lock',
     });
